bookstore/requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
crispy-bootstrap5==2025.6
cryptography==45.0.4
Django==5.2.3
django-allauth==65.9.0
django-crispy-forms==2.4
django-environ==0.12.0
idna==3.10
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pycparser==2.22
PyJWT==2.10.1
requests==2.32.4
requests-oauthlib==2.0.0
sqlparse==0.5.3
urllib3==2.5.0
